+++ src/getHtml.ts
@@ -17,7 +17,7 @@ export function getResourcePath(webview: vscode.Webview, context: vscode.Extensi
  * creates the html for the csv editor
  * @param context 
  */
-export function createEditorHtml(webview: vscode.Webview, context: vscode.ExtensionContext): string {
+export function createEditorHtml(webview: vscode.Webview, context: vscode.ExtensionContext, initialVars: InitialVars): string {
 
 	const _getResourcePath = getResourcePath.bind(undefined, webview, context)
 
@@ -209,6 +209,11 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 											</div>
 										</div>
 
+										<!-- is also works when the file is not in the workspace but opened (visible) in the vs code editor but that's too special -->
+										<span id="source-file-unwatched-indicator" class="hoverable tooltip op-hidden is-tooltip-left is-tooltip-multiline" style="float: right;margin-right: 5px;"
+											data-tooltip="The source file cannot be watched. Changes to it will not notify this view! This happens if you edit a file not inside your workspace. Actually it works when the file is not in your workspace and you ONLY edit the file inside vs code.">
+											<i class="fas fa-eye"></i>
+										</span>
 										<span id="unsaved-changes-indicator" class="hoverable unsaved-changes-indicator op-hidden tooltip is-tooltip-left" style="float: right;margin-right: 5px;"
 											data-tooltip="You might have unsaved changes">
 											<i class="fas fa-save"></i>
@@ -507,7 +512,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 	let helpModalHtml = ``
 	{
 		helpModalHtml = `
-		<div id="help-modal" class="modal my-modal">
+		<div id="help-modal" class="modal">
 		<div class="modal-background"></div>
 		<div class="modal-content">
 			<div class="box">
@@ -585,7 +590,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 	let askReadAgainModalHtml = ``
 	{
 		askReadAgainModalHtml = `
-		<div id="ask-read-again-modal" class="modal my-modal centered">
+		<div id="ask-read-again-modal" class="modal modal-centered">
 		<div class="modal-background"></div>
 		<div class="modal-content">
 			<div class="box">
@@ -617,7 +622,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 	let askReloadFileModalHtml = ``
 	{
 		askReloadFileModalHtml =`
-		<div id="ask-reload-file-modal" class="modal my-modal centered">
+		<div id="ask-reload-file-modal" class="modal modal-centered">
 		<div class="modal-background"></div>
 		<div class="modal-content">
 			<div class="box">
@@ -626,7 +631,8 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 				<p>
 					Are you sure you want to read the source file again? <br />
 					All changes to the table will be discarded! <br />
-					This will also update the snapshot of the file that is used for the reset data feature.
+					<br />
+					<i>This will also update the snapshot of the file that is used for the reset data feature.</i>
 				</p>
 
 				<div style="margin-top: 1em">
@@ -642,6 +648,40 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 			</div>
 		</div>
 		<button class="modal-close is-large" aria-label="close" onclick="toggleAskReloadFileModalDiv(false)"></button>
+	</div>
+		`
+	}
+
+	let sourceFileChangedModalHtml = ``
+	{
+		sourceFileChangedModalHtml = `
+		<div id="source-file-changed-modal" class="modal modal-centered">
+		<div class="modal-background"></div>
+		<div class="modal-content">
+			<div class="box">
+				<h3 class="title is-3">Source file changed</h3>
+
+				<p>
+					The source file changed, thus the tabel is not up-to-date. <br />
+					You can reload the file content which will discard all changes to the table! <br />
+					Or you can ignore the changes. <br />
+					<br />
+					<i>This will also update the snapshot of the file that is used for the reset data feature.</i>
+				</p>
+
+				<div style="margin-top: 1em">
+					<button class="button is-warning" onclick="reloadFileFromDisk()">
+						<span>Reload</span>
+					</button>
+
+					<button style="margin-left: 0.5em" class="button is-outlined" onclick="toggleSourceFileChangedModalDiv(false)">
+						<span>Cancel</span>
+					</button>
+				</div>
+
+			</div>
+		</div>
+		<button class="modal-close is-large" aria-label="close" onclick="toggleSourceFileChangedModalDiv(false)"></button>
 	</div>
 		`
 	}
@@ -688,8 +728,11 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 
 	${askReloadFileModalHtml}
 
+	${sourceFileChangedModalHtml}
+
 	<script>
 	var initialConfig = ${JSON.stringify(config)};
+	var initialVars = ${JSON.stringify(initialVars)};
 	</script>
 
 	 <script src="${progressJs}"></script>